한솔_qna/board 주소정리

+++ src/main/java/proj21_funding/controller/QNAController.java
@@ -37,7 +37,7 @@ public class QNAController {
 	@Autowired
 	MessageService userService;
 
-	@RequestMapping("/servicecenter/all_view")
+	@RequestMapping("/qnaallview")
 	public ModelAndView qnaAll() {
 		List<QNA> qna = qnaService.showQNAByUserId(1);
 		List<BoardCategory> bc = bcService.showBCByClass("qna");
@@ -48,7 +48,7 @@ public class QNAController {
 		return mav;
 	}
 
-	@RequestMapping("/servicecenter/admin_view")
+	@RequestMapping("/qnaadminview")
 	public ModelAndView qnaAllAdmin(
 			@RequestParam(value = "currentPage", required = false, defaultValue = "1") int currentPage,
 			@RequestParam(value = "cntPerPage", required = false, defaultValue = "10") int cntPerPage,
@@ -56,20 +56,23 @@ public class QNAController {
 			HttpServletResponse response) throws Exception {
 		UserAuthInfo user = (UserAuthInfo) session.getAttribute("authInfo");
 
+		ModelAndView mav = new ModelAndView();
+
 		response.setContentType("text/html;charset=utf-8");
 		PrintWriter out = response.getWriter();
 
-		if (user.getUserNo() <= 0 || session.getAttribute("authInfo") == null) {
+		if (user.getUserNo() >= 0 || session.getAttribute("authInfo") == null) {
 			out.println("<script type='text/javascript'>");
 			out.println("alert('잘못된 접근입니다.');");
 			out.println("history.back();");
 			out.println("</script>");
 			out.flush();
+			
+			mav.setViewName("servicecenter/all_view");
+			return mav;
 		}
 
 		List<BoardCategory> bc = bcService.showBCByClass("qna");
-
-		ModelAndView mav = new ModelAndView();
 
 		int listCnt = qnaService.QNACount();
 		Pagination pagination = new Pagination(currentPage, cntPerPage, pageSize);
@@ -83,7 +86,7 @@ public class QNAController {
 		return mav;
 	}
 
-	@RequestMapping("/servicecenter/user_view")
+	@RequestMapping("/qnauserview")
 	public ModelAndView qnaUser(
 			@RequestParam(value = "currentPage", required = false, defaultValue = "1") int currentPage,
 			@RequestParam(value = "cntPerPage", required = false, defaultValue = "10") int cntPerPage,
@@ -108,7 +111,7 @@ public class QNAController {
 		return mav;
 	}
 
-	@RequestMapping("/servicecenter/view_detail/{qnaNo}")
+	@RequestMapping("/qnadetail/{qnaNo}")
 	public ModelAndView detail(@PathVariable("qnaNo") int qnaNo) {
 		List<BoardCategory> bc = bcService.showBCByClass("qna");
 		QNA qna = qnaService.showQNAByNo(qnaNo);
@@ -120,7 +123,7 @@ public class QNAController {
 		return mav;
 	}
 
-	@RequestMapping("/servicecenter/user_write")
+	@RequestMapping("/qnawrite")
 	public ModelAndView WriteAll() {
 		List<BoardCategory> bc = bcService.showBCByClass("qna");
 		ModelAndView mav = new ModelAndView();
@@ -129,28 +132,34 @@ public class QNAController {
 		return mav;
 	}
 
-	@RequestMapping("/servicecenter/amdin_reply/{qnaNo}")
+	@RequestMapping("/qnareply/{qnaNo}")
 	public ModelAndView ReplyAll(@PathVariable("qnaNo") int qnaNo, HttpSession session, HttpServletResponse response)
 			throws IOException {
 		QNA qna = qnaService.showQNAByNo(qnaNo);
 		List<BoardCategory> bc = bcService.showBCByClass("qna");
 		UserInfo user = userService.showUserbyNo(qnaNo);
-		UserInfo admin = (UserInfo) session.getAttribute("authInfo");
+		UserAuthInfo admin = (UserAuthInfo) session.getAttribute("authInfo");
 
+		ModelAndView mav = new ModelAndView();
+		
 		response.setContentType("text/html;charset=utf-8");
 		PrintWriter out = response.getWriter();
 
-		if (admin.getUserNo() <= 0 || session.getAttribute("authInfo") == null) {
+		if (admin.getUserNo() >= 0 || session.getAttribute("authInfo") == null) {
 			out.println("<script type='text/javascript'>");
 			out.println("alert('잘못된 접근입니다.');");
 			out.println("history.back();");
 			out.println("</script>");
 			out.flush();
+			
+			mav.setViewName("servicecenter/all_view");
+			return mav;
 		}
 
+		mav.setViewName("servicecenter/admin_reply");
 		System.out.println(qna);
 		System.out.println(user);
-		ModelAndView mav = new ModelAndView("servicecenter/amdin_reply", "qna", qna);
+		mav.addObject("qna", qna);
 		mav.addObject("user", user);
 		mav.addObject("bc", bc);
 		return mav;
@@ -172,10 +181,27 @@ public class QNAController {
 	}
 
 	@PostMapping("/qnaRpSuccess")
-	public ModelAndView qnaReply(QNA qna) {
+	public ModelAndView qnaReply(QNA qna, HttpSession session, HttpServletResponse response) {
 		try {
 			System.out.println(qna);
+
+			UserAuthInfo admin = (UserAuthInfo) session.getAttribute("authInfo");
+
+			response.setContentType("text/html;charset=utf-8");
+			PrintWriter out = response.getWriter();
+			
+			if (admin.getUserNo() >= 0 || session.getAttribute("authInfo") == null) {
+				out.println("<script type='text/javascript'>");
+				out.println("alert('잘못된 접근입니다.');");
+				out.println("history.back();");
+				out.println("</script>");
+				out.flush();
+
+				return detail(qna.getQnaNo());
+			}
+
 			qnaService.replyQNA(qna);
+			
 			return detail(qna.getQnaNo());
 
 		} catch (Exception e) {
@@ -184,11 +210,33 @@ public class QNAController {
 		}
 	}
 
-	@GetMapping("/delete/{qnaNo}")
-	public String deleteSuccess(@PathVariable("qnaNo") int qnaNo) {
+	@GetMapping("/qnadelete/{qnaNo}")
+	public String deleteSuccess(@PathVariable("qnaNo") int qnaNo, HttpSession session, HttpServletResponse response) {
 		try {
 			System.out.println(qnaNo);
+			QNA qna = qnaService.showQNAByNo(qnaNo);
+			
+			UserAuthInfo user = (UserAuthInfo) session.getAttribute("authInfo");
+			UserInfo resUser = userService.showUserbyNo(qna.getUserNo().getUserNo());
+			
+			response.setContentType("text/html;charset=utf-8");
+			PrintWriter out = response.getWriter();
+			
+			System.out.println(user.getUserNo());
+			System.out.println(resUser.getUserNo());
+			
+			if (user.getUserNo() != resUser.getUserNo()) {
+				out.println("<script type='text/javascript'>");
+				out.println("alert('잘못된 접근입니다.');");
+				out.println("history.back();");
+				out.println("</script>");
+				out.flush();
+
+				return "/";
+			}
+
 			qnaService.removeQNA(qnaNo);
+			
 			return "servicecenter/delete";
 
 		} catch (Exception e) {
